Audit: widen runtime rule storage to full PWM range

+++ FuzzyController_src/FB_FuzzyConfigManager.h
@@ -29,9 +29,6 @@ typedef struct
     float Peak;
 } FuzzyMFConfig_t;
 
-/* Runtime compact scaling representation used by the configuration manager.
- * This is intentionally distinct from FuzzyScalingConfig_t in FB_FuzzyScaling.h.
- */
 typedef struct
 {
     float errorScale;
@@ -51,7 +48,7 @@ typedef struct
     uint16_t version;
     bool enable;
     FuzzyMFConfig_t MF[7];
-    uint8_t rule[7][7];
+    int16_t rule[7][7];
     FuzzyRuntimeScalingConfig_t scaling;
     FuzzyFFConfig_t ff[16];
     uint8_t ffSize;
@@ -67,7 +64,7 @@ MY_API void FB_FuzzyConfig_Init(FB_FuzzyConfigManager_t *fb);
 MY_API void FB_FuzzyConfig_LoadDefault(FB_FuzzyConfigManager_t *fb);
 MY_API bool FB_FuzzyConfig_Check(FB_FuzzyConfigManager_t *fb);
 MY_API bool FB_FuzzyConfig_Apply(FB_FuzzyConfigManager_t *cfg, FB_FuzzyController_t *controller);
-MY_API bool FB_FuzzyConfig_SetRule(FB_FuzzyConfigManager_t *fb, uint8_t e, uint8_t de, uint8_t output);
+MY_API bool FB_FuzzyConfig_SetRule(FB_FuzzyConfigManager_t *fb, uint8_t e, uint8_t de, int16_t output);
 MY_API bool FB_FuzzyConfig_SetMF(FB_FuzzyConfigManager_t *fb, uint8_t index, FuzzyMFConfig_t *mf);
 
 //------------------------------------------------------------------------------------//
